requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.3
django-material==1.0.0
django-oauth-toolkit==1.0.0
idna==2.5
MySQL-python==1.2.5
oauthlib==2.0.2
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
